The Code
This code works like this.  You either choose to encrypt or decrypt a
message, then you input a key, which is how many characters you shift,
then input a message, then go through each character in the message and
shift it key amount of times.  If it goes greater that 26 then the
modulus wraps around.  Then it stores it where it got the original
character, and prints the final message